refactor: split interaction into multiple files
