+++ packages/mask/popups/pages/Wallet/Interaction/WatchTokenRequest.tsx
@@ -2,16 +2,17 @@ import { memo } from 'react'
 import { Box, Typography } from '@mui/material'
 import { makeStyles } from '@masknet/theme'
 import { NetworkPluginID } from '@masknet/shared-base'
-import { useChainContext, useWallet, useWeb3Hub } from '@masknet/web3-hooks-base'
-import type { TransactionDetail } from '../../pages/Wallet/type.js'
-import { useMaskSharedTrans } from '../../../shared-ui/index.js'
+import { useChainContext, useWallet, useWeb3Hub, useWeb3State } from '@masknet/web3-hooks-base'
+import { useMaskSharedTrans } from '../../../../shared-ui/index.js'
 import { TokenIcon } from '@masknet/shared'
-import { type ChainId } from '@masknet/web3-shared-evm'
-import { TokenDetailUI } from '../../pages/Wallet/TokenDetail/index.js'
-import { ContractSection } from '../../../../shared/src/UI/components/CoinMetadataTable/ContractSection.js'
-import { useAsset } from '../../pages/Wallet/hooks/useAsset.js'
-import { CollectibleDetailUI } from '../../pages/Wallet/CollectibleDetail/index.js'
+import { SchemaType, type ChainId } from '@masknet/web3-shared-evm'
+import { TokenDetailUI } from '../TokenDetail/index.js'
+import { ContractSection } from '../../../../../shared/src/UI/components/CoinMetadataTable/ContractSection.js'
+import { useAsset } from '../hooks/useAsset.js'
+import { CollectibleDetailUI } from '../CollectibleDetail/index.js'
 import { useAsync } from 'react-use'
+import type { InteractionItemProps } from './interaction.js'
+import { TokenType } from '@masknet/web3-shared-base'
 
 const useStyles = makeStyles()({
     title: { fontSize: 28 },
@@ -21,20 +22,51 @@ const useStyles = makeStyles()({
     iconContainer: { display: 'flex', alignItems: 'center', gap: 12 },
 })
 
-interface WatchTokenRequestProps {
-    transaction: TransactionDetail
-}
-
 // TODO: wallet_watchAsset SHOULD check the name and symbol fields, and the contract address and chainId against a list of well-known tokens. If the name and/or symbol are similar to ones on the list but the chainId/address don’t match, a warning SHOULD be presented to the user.
-export const WatchTokenRequest = memo<WatchTokenRequestProps>(function WatchTokenRequest({ transaction }) {
+export const WatchTokenRequest = memo<InteractionItemProps>((props) => {
+    const { currentRequest: request, setConfirmAction } = props
     const t = useMaskSharedTrans()
     const { classes } = useStyles()
+    const { Message, Token } = useWeb3State(NetworkPluginID.PLUGIN_EVM)
+    const wallet = useWallet()
+
+    const { params } = request.request.arguments
+    setConfirmAction(async () => {
+        const type = params[0].type
+        const address = params[0].options.address
+        if (type === 'ERC21') {
+            // TODO: custom name currently are ignored
+            await Token!.addToken!(wallet!.address, {
+                address,
+                chainId,
+                schema: SchemaType.ERC20,
+                type: TokenType.Fungible,
+                id: `${chainId}.${address}`,
+                isCustomToken: true,
+            })
+        } else if (type === 'ERC721' || type === 'ERC1155') {
+            const { tokenId, symbol, name = 'NFT' } = params[0].options
+            const schema = type === 'ERC21' ? SchemaType.ERC721 : SchemaType.ERC1155
+            await Token!.addNonFungibleTokens!(wallet!.address, { address, chainId, name, schema, symbol }, [tokenId])
+            await Token!.addToken!(wallet!.address, {
+                id: `${chainId}.${address}.${tokenId}`,
+                chainId,
+                tokenId,
+                type: TokenType.NonFungible,
+                schema,
+                address,
+                isCustomToken: true,
+            })
+        }
+        // It is "deny" because we don't want send it to the upstream RPC.
+        await Message!.denyRequest(request.ID)
+    })
 
     const { chainId } = useChainContext<NetworkPluginID.PLUGIN_EVM>()
     const {
         options: { address, symbol, image, tokenId },
         type,
-    } = transaction.payload.params![0] as {
+    } = params[0] as {
         options: { address: string; symbol: string; image: string; tokenId?: string }
         type: 'ERC20' | 'ERC721' | 'ERC1155'
     }
@@ -96,6 +128,8 @@ export const WatchTokenRequest = memo<WatchTokenRequestProps>(function WatchToke
         </Box>
     )
 })
+WatchTokenRequest.displayName = 'WatchTokenRequest'
+
 function NonFungibleAsset({ address, chainId, tokenId }: { address: string; chainId: ChainId; tokenId: string }) {
     const t = useMaskSharedTrans()
     const { classes } = useStyles()
