+++ packages/mask/popups/pages/Wallet/Interaction/interaction.tsx
@@ -1,0 +1,221 @@
+import { type ReasonableMessage } from '@masknet/web3-shared-base'
+import { EthereumMethodType, type MessageRequest } from '@masknet/web3-shared-evm'
+import React, { memo, startTransition, useCallback, useRef, useState } from 'react'
+import type { JsonRpcResponse } from 'web3-core-helpers'
+import { WatchTokenRequest } from './WatchTokenRequest.js'
+import { WalletSignRequest } from './WalletSignRequest.js'
+import { TransactionRequest } from './TransactionRequest.js'
+import { Box, Dialog, DialogActions, DialogContent, DialogContentText, IconButton, Typography } from '@mui/material'
+import { BottomController } from '../../../components/BottomController/index.js'
+import { ActionButton, makeStyles, usePopupCustomSnackbar } from '@masknet/theme'
+import { useMaskSharedTrans } from '../../../../shared-ui/index.js'
+import { Icons } from '@masknet/icons'
+import { useAsyncFn } from 'react-use'
+import { useWeb3State } from '@masknet/web3-hooks-base'
+import { NetworkPluginID, PopupRoutes } from '@masknet/shared-base'
+import Services from '#services'
+import { useNavigate } from 'react-router-dom'
+import { WalletAssetTabs } from '../type.js'
+import urlcat from 'urlcat'
+
+const useStyles = makeStyles()({
+    left: {
+        transform: 'rotate(90deg)',
+    },
+    right: {
+        transform: 'rotate(-90deg)',
+    },
+    text: {
+        fontSize: 12,
+        fontWeight: 700,
+        lineHeight: '16px',
+    },
+})
+
+interface InteractionProps {
+    currentRequest: ReasonableMessage<MessageRequest, JsonRpcResponse>
+    totalMessages: number
+    currentMessageIndex: number
+    setMessageIndex(count: number): void
+
+    paymentToken: string
+    setPaymentToken: (paymentToken: string) => void
+}
+
+export const Interaction = memo((props: InteractionProps) => {
+    const { currentRequest } = props
+    const t = useMaskSharedTrans()
+    const navigate = useNavigate()
+    const { Message } = useWeb3State(NetworkPluginID.PLUGIN_EVM)
+    const { showSnackbar } = usePopupCustomSnackbar()
+
+    const [isDangerRequest, setIsDanger] = useState(false)
+    const [dangerDialogOpen, setDangerDialogOpen] = useState(false)
+    const [confirmVerb, setConfirmVerb] = useState(t.confirm())
+    const confirmAction = useRef<(lastRequest: boolean) => Promise<void>>(async () => {})
+
+    const [{ loading: cancelLoading }, onCancel] = useAsyncFn(async () => {
+        await Message!.denyRequest(currentRequest.ID)
+        if (currentRequest.origin) await Services.Helper.removePopupWindow()
+        navigate(PopupRoutes.Wallet, { replace: true })
+    }, [currentRequest.ID, Message, currentRequest.origin])
+
+    const isLastRequest = props.totalMessages === 1
+    const [{ loading: confirmLoading }, onConfirm] = useAsyncFn(async () => {
+        try {
+            await confirmAction.current(isLastRequest)
+
+            // packages/web3-providers/src/Web3/EVM/state/Message.ts denies all requests after approving one
+            // TODO: requests that is not a transaction should not be denied
+            if (currentRequest.origin) await Services.Helper.removePopupWindow()
+            else navigate(urlcat(PopupRoutes.Wallet, { tab: WalletAssetTabs.Activity }), { replace: true })
+        } catch (error) {
+            showSnackbar(
+                <Typography textAlign="center" width="275px">
+                    {t.popups_wallet_rpc_error()}
+                    <br />
+                    {String((error as any).message)}
+                </Typography>,
+                { variant: 'error', autoHideDuration: 5000 },
+            )
+        }
+    }, [confirmAction.current, isLastRequest])
+
+    const actionRunning = confirmLoading || cancelLoading
+    const CancelButton = (
+        <ActionButton
+            loading={cancelLoading}
+            disabled={actionRunning}
+            onClick={() => {
+                if (isDangerRequest && dangerDialogOpen) setDangerDialogOpen(false)
+                else onCancel()
+            }}
+            fullWidth
+            variant="outlined">
+            {t.cancel()}
+        </ActionButton>
+    )
+    const ConfirmButton = (
+        <ActionButton
+            loading={confirmLoading}
+            disabled={actionRunning}
+            sx={isDangerRequest ? { background: (theme) => theme.palette.maskColor.danger } : undefined}
+            onClick={() => {
+                if (isDangerRequest && !dangerDialogOpen) return setDangerDialogOpen(true)
+                else onConfirm()
+            }}
+            fullWidth>
+            {confirmVerb}
+        </ActionButton>
+    )
+
+    return (
+        <Box flex={1} display="flex" flexDirection="column">
+            <Box p={2} display="flex" flexDirection="column" flex={1} maxHeight="calc(100vh - 142px)" overflow="auto">
+                <InteractionItem
+                    paymentToken={props.paymentToken}
+                    setPaymentToken={props.setPaymentToken}
+                    currentRequest={currentRequest}
+                    setConfirmVerb={setConfirmVerb}
+                    setIsDanger={setIsDanger}
+                    setConfirmAction={useCallback((f) => (confirmAction.current = f), [])}
+                />
+                <Pager {...props} />
+            </Box>
+            {dangerDialogOpen ?
+                <DangerDialog cancel={ConfirmButton} confirm={CancelButton} />
+            :   null}
+            <BottomController>
+                {CancelButton}
+                {ConfirmButton}
+            </BottomController>
+        </Box>
+    )
+})
+Interaction.displayName = 'Interaction'
+
+export interface InteractionItemProps {
+    currentRequest: ReasonableMessage<MessageRequest, JsonRpcResponse>
+    setIsDanger(isDanger: boolean): void
+    setConfirmVerb(verb: string): void
+    setConfirmAction(action: (isLastRequest: boolean) => Promise<void>): void
+
+    // transaction only
+    paymentToken: string
+    setPaymentToken: (paymentToken: string) => void
+}
+const InteractionItem = memo((props: InteractionItemProps) => {
+    switch (props.currentRequest.request.arguments.method) {
+        case EthereumMethodType.wallet_watchAsset:
+            return <WatchTokenRequest {...props} />
+        case EthereumMethodType.eth_sign:
+        case EthereumMethodType.eth_signTypedData_v4:
+        case EthereumMethodType.personal_sign:
+            return <WalletSignRequest {...props} />
+        default:
+            return <TransactionRequest {...props} />
+    }
+})
+InteractionItem.displayName = 'InteractionItem'
+
+const Pager = memo((props: InteractionProps) => {
+    const { currentMessageIndex, currentRequest, setMessageIndex, totalMessages } = props
+    const { classes } = useStyles()
+    const t = useMaskSharedTrans()
+    const navigate = useNavigate()
+    const { Message } = useWeb3State(NetworkPluginID.PLUGIN_EVM)
+    const [{ loading: cancelAllLoading }, handleCancelAllRequest] = useAsyncFn(async () => {
+        await Message!.denyAllRequests()
+        if (currentRequest.origin) await Services.Helper.removePopupWindow()
+        else navigate(PopupRoutes.Wallet, { replace: true })
+    }, [Message, currentRequest.origin])
+
+    if (totalMessages <= 1) return null
+    return (
+        <Box display="flex" flexDirection="column" alignItems="center" marginTop="auto">
+            <Box display="flex" alignItems="center">
+                <IconButton
+                    disabled={currentMessageIndex === 0}
+                    onClick={() => startTransition(() => setMessageIndex(currentMessageIndex - 1))}>
+                    <Icons.ArrowDrop size={16} className={classes.left} />
+                </IconButton>
+                <Typography className={classes.text}>
+                    {t.popups_wallet_multiple_requests({
+                        index: String(currentMessageIndex + 1),
+                        total: String(totalMessages),
+                    })}
+                </Typography>
+                <IconButton
+                    onClick={() => startTransition(() => setMessageIndex(currentMessageIndex + 1))}
+                    disabled={currentMessageIndex === totalMessages - 1}>
+                    <Icons.ArrowDrop size={16} className={classes.right} />
+                </IconButton>
+            </Box>
+
+            <ActionButton variant="text" color="info" onClick={handleCancelAllRequest} loading={cancelAllLoading}>
+                {t.popups_wallet_reject_all_requests({ total: String(totalMessages) })}
+            </ActionButton>
+        </Box>
+    )
+})
+Pager.displayName = 'Pager'
+
+function DangerDialog({ cancel, confirm }: Record<'cancel' | 'confirm', React.ReactNode>) {
+    const t = useMaskSharedTrans()
+    return (
+        <Dialog open>
+            <DialogContent>
+                <DialogContentText variant="overline">
+                    {t.popups_wallet_sign_in_danger_confirm_title()}
+                </DialogContentText>
+                <DialogContentText color={(theme) => theme.palette.maskColor.danger}>
+                    {t.popups_wallet_sign_in_danger_confirm()}
+                </DialogContentText>
+            </DialogContent>
+            <DialogActions>
+                {cancel}
+                {confirm}
+            </DialogActions>
+        </Dialog>
+    )
+}
