+++ packages/mask/popups/pages/Wallet/index.tsx
@@ -22,7 +22,7 @@ const AddToken = lazy(() => import('./AddToken/index.js'))
 const GasSetting = lazy(() => import(/* webpackMode: 'eager' */ './GasSetting/index.js'))
 const Transfer = lazy(() => import(/* webpackMode: 'eager' */ './Transfer/index.js'))
 const ContactList = lazy(() => import(/* webpackMode: 'eager' */ './ContactList/index.js'))
-const ContractInteraction = lazy(() => import(/* webpackMode: 'eager' */ './Interaction/index.js'))
+const ContractInteraction = lazy(() => import(/* webpackMode: 'eager' */ './Interaction/page.js'))
 const ResetWallet = lazy(() => import('./ResetWallet/index.js'))
 const SetPaymentPassword = lazy(() => import('./SetPaymentPassword/index.js'))
 const ChangeOwner = lazy(() => import('./ChangeOwner/index.js'))
